remove unnecessary files, clean up index.js
