+++ src/index.js
@@ -1,157 +1,305 @@
-const { Client } = require('@elastic/elasticsearch')
-const fs = require('fs')
+const {
+  Client
+} = require('@elastic/elasticsearch');
+const fetch = require('node-fetch');
 
+class Reader {
+  /**
+   * reads in data, whether it be from a file or
+   * an API endpoint
+   */
 
-const client = new Client({
-  node: 'http://localhost:9200',
-  auth: {
-    username: 'elastic',
-    password: process.env.ES_PW ? process.env.ES_PW: "Dpm#1217"
-  }
-})
-
-client.cluster.health({},function(err,resp,status) {
-  if (err) {
-    console.error(err)
-  }
-});
-
-class AlgoLogger {
-  /* using file name, read in algo logs and bulk upload them to elasticsearch */
-  constructor(fPath, elasticsearchClient, algoUrl) {
-    this.fPath = fPath;
-    this.elasticsearchClient = elasticsearchClient;
-    this.algoUrl= algoUrl
-  }
-
-  async parse() {
-    // parse log files, split into array where each item is one line
-    // return data parsed and structured for elasticsearch
-
-    await fs.readFile(this.fPath, 'utf8', async (err, data) => {
-      if (err) { console.error(err) } else {
-        let unstructuredDataArr = data.split('\n');
-        let structuredData = await this.structure(unstructuredDataArr);
-        this.data = structuredData;
-
-        return this.data;
-      }
-    })
+  constructor(url, keysToFetch = null, {
+    headers,
+    params,
+    body
+  } = {}) {
+    /**
+     * `location` -> (String): the location to read the
+     * data from
+     */
+    this.url = url
+    this.keysToFetch = keysToFetch
+    this.headers = headers;
+    this.params = params
+    this.body = body
   }
 
-  async structure(unstructuredDataArr) {
-    // structure data in log files for es
+  async read() {
+    /**
+     * call the pingEndpoint helper function and return
+     * Parser(this)
+     */
+
+    let response;
+    response = await this._pingEndpoint(this.url)
+
+    return new Parser(this);
+  }
+
+  async _pingEndpoint() {
+    /**
+     * ping `this.url` and return JSON data for top-transactions.
+     */
+
+    let json;
+
+    try {
+      let res = await fetch(this.url, {
+        headers: this.headers
+      })
+      json = await res.json();
+    } catch (err) {
+      return new Error(err);
+    }
+
+    this.data = json['top-transactions'];
+  }
+
+}
+
+class Parser {
+  constructor(reader) {
+    this.reader = reader
+  }
+
+  async structure() {
+    /**
+     * structure data in the log files for
+     * elasticsearch
+     * unstructuredData -> (Array): array containing raw line items
+     *                                 from the log files
+     */
+
     let dataForElasticsearch = new Array()
 
-    for (var lineNum in unstructuredDataArr) {
-      const line = unstructuredDataArr[lineNum]
-      const structuredLine = await this._buildElasticsearchPayloadFrom(line, lineNum)
-      console.log(structuredLine);
+    for (var i in this.reader.data) {
+      const line = this.reader.data[i]
+      const structuredLine = await this._buildElasticsearchPayloadFrom(line, i)
       dataForElasticsearch.push(
         structuredLine
       );
     }
 
-    return dataForElasticsearch;
+    this.structuredData = dataForElasticsearch;
+    return new Uploader(this)
   }
 
-  async _buildElasticsearchPayloadFrom(log, logLineNum) {
-    /* for a given log line, structure it's payload */
-    let keysToFetch = new Array("Context", "Hash", "Round", "Sender", "Type", "msg");
-    let lineData = new Object()
+  async _buildElasticsearchPayloadFrom(rawData, i) {
+    /**
+     * for a given log line, extract the needed data and format an Object
+     *
+     * rawData -> raw data from file or API endpoint
+     * i -> (integer): the ID for the elasticsearch document
+     *
+     * returns -> (Array): cleaned line data
+     */
+    let keysToFetch = this.reader.keysToFetch !== null ? this.reader.keysToFetch : Object.keys(rawData['txn']);
+    let structuredData = new Object()
+    let jsonifiedData;
 
-    let jsonifiedLog;
-    try {
-      jsonifiedLog = JSON.parse(log);
+    if (typeof rawData !== 'object') {
+      try {
+        jsonifiedData = JSON.parse(rawData);
+      } catch (e) {
+        return
+      } // if cannot be parsed by JSON, return (skip the line)
+    } else {
+      jsonifiedData = rawData;
     }
-    catch(e) { return } // if cannot be parsed by JSON, return
 
-    keysToFetch.forEach( (key) => {
-      lineData[key.toLowerCase()] = jsonifiedLog[key]
+    keysToFetch.forEach((key) => {
+      structuredData[key.toLowerCase()] = jsonifiedData['txn'][key]
+      structuredData['sig'] = jsonifiedData['sig']
     })
 
-    lineData["date"] = new Date()
-    lineData["id"] = logLineNum
-    return lineData
+    structuredData["date"] = new Date()
+    structuredData["id"] = i
+    return structuredData
+  }
+
+
+}
+
+
+class Uploader {
+
+  constructor(parser) {
+    this.parser = parser;
   }
 
   async _indiceExistsFor(indexName) {
-    /* check if index `indexName` exists, return result */
+    /**
+     * check if `indexName` exists as an elasticsearch index.
+     * indexName -> (String): the name of an index
+     */
 
-    const res = await client.indices.exists({index: indexName})
+    const res = await this.client.indices.exists({
+      index: indexName
+    })
     const exists = res.body
 
     return exists
   }
 
   async _createIndex(indexName, indexMappings) {
-    // create elasticsearch index
-    await client.indices.create({
+    /**
+     * create an elasticsearch index based on `indexName` and `indexMappings`
+     *
+     * indexName -> (String): the name of the index to create
+     * indexMappings -> (Object): the mappings of the data types for the index
+     */
+    await this.client.indices.create({
       index: indexName,
       body: {
         mappings: {
           properties: indexMappings
         }
       }
-    }, { ignore: [400] })
+    }, {
+      ignore: [400]
+    })
   }
 
   async _uploadData(indexName) {
+    /**
+     * upload data to the elasticsearch index.
+     *
+     * indexName -> (String): the name of the index
+     */
 
-    const body = this.data.flatMap( (doc) => [{ index: { _index: indexName } }, doc])
-    const { body: bulkResponse } = await client.bulk({ refresh: true, body })
+    const body = this.parser.structuredData.flatMap((doc) => [{
+      index: {
+        _index: indexName
+      }
+    }, doc])
+    const {
+      body: bulkResponse
+    } = await this.client.bulk({
+      refresh: true,
+      body
+    })
     return bulkResponse
   }
 
-  async upload(indexName, indexMappings) {
-    // upload to elasticsearch
+  async uploadTo(client, {
+    indexName,
+    mappings
+  }) {
+    /**
+     * upload data to elasticsearch
+     *
+     * indexName -> (String): the name of the index
+     * indexMappings -> (Object): the data type mappings for the fields in the index
+     */
+    this.client = client;
+
+    console.log(indexName);
+
     let indexExists = await this._indiceExistsFor(indexName);
 
     if (!indexExists) {
       // if indice does not exist...
       console.log("index does not exist, creating it now");
-      this._createIndex(indexName, indexMappings);
+      this._createIndex(indexName, mappings);
     }
 
     // index exists, we can now upload data to it.
     await this._uploadData(indexName)
-    let {body: { count: count } } = await client.count({ index: indexName })
+    let {
+      body: {
+        count: count
+      }
+    } = await this.client.count({
+      index: indexName
+    })
 
     console.log(`There are ${count} total documents in index ${indexName}`);
   }
 
   async testIndex(indexName, queryParams) {
-    const { body } = await client.search({
-        index: indexName,
-        body: queryParams
-      })
+    /**
+     * helper function to test various queries on the elasticsearch index
+     *
+     * indexName -> (String): the name of the index
+     * queryParams -> (Object): the queries used to search elasticsearch
+     */
+    const {
+      body
+    } = await this.client.search({
+      index: indexName,
+      body: queryParams
+    })
 
-      console.log(body.hits.hits);
+    console.log(body.hits.hits);
   }
-
 }
 
-const Algo = new AlgoLogger('/Users/danielmurphy/Desktop/ELK-to-Algo/node.log', '', 'sdfgfsh');
-// Algo.parse().catch(console.log);
-// let mappings = {
-//   context: {type: "text"},
-//   hash: {type: "text"},
-//   round: {type: "integer"},
-//   sender: {type: "text"},
-//   type: {type: "text"},
-//   msg: {type: "text"},
-//   date: {type: "date"},
-//   id: {type: "integer"}
-// }
-// Promise.resolve(Algo.upload("algorand", mappings).catch(console.log)).then( (res) => {console.log(res);});
-// Algo.testIndex("algorand",
-//   {
-//     query: {
-//       match: {
-//         type: "VoteAccepted"
-//       }
-//     }
-//   }).catch(console.log)
+const elasticsearchClient = new Client({
+  node: 'http://localhost:9200',
+  auth: {
+    username: 'elastic',
+    password: process.env.ES_PW
+  }
+})
+
+let mappings = {
+  arcv: {
+    type: "text"
+  },
+  sig: {
+    type: "text"
+  },
+  fee: {
+    type: "integer"
+  },
+  fv: {
+    type: "integer"
+  },
+  gen: {
+    type: "text"
+  },
+  gh: {
+    type: "text"
+  },
+  lv: {
+    type: "integer"
+  },
+  note: {
+    type: "text"
+  },
+  snd: {
+    type: "text"
+  },
+  type: {
+    type: "axfer"
+  },
+  xaid: {
+    type: "integer"
+  },
+  date: {
+    type: "date"
+  },
+  id: {
+    type: "text"
+  }
+}
+
+const reader = new Reader("http://127.0.0.1:8080/v2/transactions/pending", null, {
+  headers: {
+    "X-Algo-API-Token": process.env.TOKEN
+  }
+})
 
 
-exports.AlgoLogger = AlgoLogger
+async function run() {
+  let parser = await reader.read();
+  let uploader = await parser.structure();
+  console.log(uploader);
+  await uploader.uploadTo(elasticsearchClient, {
+    mappings: mappings,
+    indexName: "algorand-final"
+  })
+}
+
+run().catch(console.log)
